// Payanak/Original/ClientApp/src/app/shared/model/contact/contact.model.ts
export class ContactModel {
  id: number;
  username: string;
  fName: string;
  lName: string;
  nickName: string;
  birthday: number;
  gender: number;
  nationalCode: string;
  telegram: string;
  instagram: string;
  specialDay: number;
  mobilePhone: string;
  latitude: string;
  longitude: string;
  address: string;
  formId: string;
  picture: string;
  credit: number;
}